Replace VirtualMethodEnumerationAlgorithm with simple GetAllMethods (#1975)
When generating generic dictionary dependencies, we need to consider the
effect of typesystem-injected methods.

Instead of calling `GetAllVirtualMethods` in addition to `GetMethods`
when generating the conditional dependencies, I decided to revisit the
whole "virtual method enumeration algorithm" business. The complexity
doesn't seem to be worth it, so I'm killing it.