Change the way we emit optional fields
Before this change, optional fields used to be represented by mutable
object nodes that created a bunch of problems: Besides the key-value
mismatches described in #1826, this was also causing e.g. garbage
unreferenced optional fields nodes getting generated (e.g. an EEType
thinks it will need an optional fields node of shape A, so we add it to
the graph and mark it. Later the EEType changes it's mind and now needs
shape B. But the shape A optional fields node has already been marked
and will be emitted into the object file.)

I'm making the optional fields node local to the EEType that needs it.
This disables the size on disk optimization that tried to reuse the
nodes. We can re-enable this later by adding a general purpose mechanism
to object writer that can collapse nodes with identical contents that
don't need to maintain identity (this can be e.g. also used to collapse
method code nodes).

Resolves #1826.