Move a few properties to GenericParameterDesc (#2107)
While working on reflection invoke stubs I had the need to define a new
child of `GenericParameterDesc`. That's when I noticed there's a lot of
boilerplate I would need to replicate. I'm moving the boilerplate to
`GenericParameterDesc`.

I'm leaving sealable things unsealed with TODOs to avoid unnecessary
breakage in the TFS-land.