Remove a layering violation in the type loader (maybe the final one?)
Type loader was calling into reflection to get metadata names for things because long time ago in a galaxy far away the type loader wasn't allowed to know about metadata. This required going over layers of indirections to get to reflection. I'm simply moving the code to the type loader.

[tfs-changeset: 1636951]