Move compilation root handling out of CompilationGroup (#2177)
Compilation group was serving two purposes - establishing the list of
roots, and controlling dependency graph expansion. This was already a
bit awkward because `MultiFileCompilationGroup` was serving two rooting
purposes - production build and consumption build - with different sets
of roots. It would get even more awkward once reflection enters the
picture.

I'm creating a new concept of a root provider. It's a class that
provides a set of roots, and you can combine them. This gives the driver
EXE a more fine grained control of what should become a root. This
potentially lifts a bunch of restrictions too (e.g. making it possible
to have two EXE files in the inputs, with one being a real entrypoint
and the other a normal AssemblyRef is just a bit of work in the driver
EXE now). It also makes it more convenient to start considering
different compilation targets ("compile to a .dll/.so").

Reflection initialization is still digusting, but at least I kicked it
out into the EXE.