Make RuntimeDeterminedMethod non-generic
I made this generic so that we can statically enforce that the canonical
method node is both a `DependencyNodeCore` and an `IMethodNode`.
Unfortunately the cast to `MethodCodeNode` doesn't work in single method
compilation mode because the method could end up being an ExternMethod.
I'm removing the genericness because it's just getting in the way.