Add support for Type.MakeArrayType/Array.CreateInstance
I thought this would require a `#if CORERT` hack in the classlib, but
turns out this  is just about emitting another mapping table. On Project
N, this mapping table only has arrays of value types in it (type loader
builds the rest on demand), but on CoreRT we don't generate enough data
structures to support type loader yet. Turns out that if we put the
reference types in the table, the type loader is able to find them and
won't try to build them.

I'm leaving a TODO to filter out the reference type arrays once we bring
up the type loader.