Fix DeveloperExperience.Console output
The WriteConsole Win32 API is documented to fail if the output is
redirected, which makes the DeveloperExperience console writing useless
for all of our automated runs. This is a regression that happened around
the time DeveloperExperience.Console was moved out to GitHub.

I won't pretend this is the right fix, but printing something with a
potentially wrong code page is stricly better than printing nothing.