Refactor ObjectNode shareability
For multi-module build mode, each node was able to consult the
ModuleCompilationGroup to decide whether it should be shared. Instead,
replace `ShouldShareNodeAcrossModules` with a new ObjectNode property
`IsShareable` which indicates whether the node should be shared if
possible. It uses local fields only, and lets the object emission layer
care about whether or not to enable sharing.