Remove NodeFactory.CompilationUnitPrefix
Remove the `CompilationUnitPrefix` static property. It was not getting
set early enough in compilation and in multi-module builds, this caused
the global tables we build to not get the prefix set (ie,
`DispatchMapTable` and friends).

Instead, ObjectWriter generates the prefix based on the output object
file name, leaving it empty in single file compilation mode.