Fix for build break caused by CS# 1639380 which introduced a new state transition in the MCG runtime state machine: by adding the new method McgModuleManager.Initialize which is explicitly called from the startup code trigger, we have introduced a state of "partial MCG initialization" before Initialize runs. In this state, some methods must not be called because they refer to the state that only gets initialized in Initialize. This was happening in the method VerifyWinRTGenericInterfaceGuids whose call graph ended up calling GetIndicesForInterface which uses one of the maps initialized during Initialize (s_runtimeTypeHandleToInterfaceIndexMap).
This simple change fixes this by removing the call to VerifyWinRTGenericInterfaceGuids from McgModule constructor and moving it to McgModuleManager.Initialize after we've initialized all the relevant maps. While I was there, I also fixed a typo in variable name I noticed in StartupCodeTrigger while looking into this issue and I expanded on a code comment in McgModule.cs. In light of the fact that this seems to be fixing general shared build break, I intend to check this in as soon as I receive any sign-off and I finish some initial testing, however I'll leave the review open such that Faizur has a chance to provide feedback especially as it's national holiday today in the US.

Thanks

Tomas

[tfs-changeset: 1639453]