Multi-module prep (#2248)
* Refactor ObjectNode shareability

For multi-module build mode, each node was able to consult the
ModuleCompilationGroup to decide whether it should be shared. Instead,
replace `ShouldShareNodeAcrossModules` with a new ObjectNode property
`IsShareable` which indicates whether the node should be shared if
possible. It uses local fields only, and lets the object emission layer
care about whether or not to enable sharing.

* Remove NodeFactory.CompilationUnitPrefix

Remove the `CompilationUnitPrefix` static property. It was not getting
set early enough in compilation and in multi-module builds, this caused
the global tables we build to not get the prefix set (ie,
`DispatchMapTable` and friends).

Instead, ObjectWriter generates the prefix based on the output object
file name, leaving it empty in single file compilation mode.