Fix a thread-unsafe issue in name mangler. (#2300)
This is exposed When the code generator compiles methods in
multithreading. The method GetMangledTypeName doesn't search the name hash
table under a lock which is potentially thread-unsafe. We don't want a
lock there for the sake of performance. However, the write to the name
hash is lock-protected. The fix here is to search first during the
protected write. If the name already exists, then do not write again.
Otherwise, that the old name and new name are two different string pointers
annoys the name hash table which is a immutable hash table.