Add a concept of "compiler generated assembly"
During the course of compilation, the compiler generates various static
methods that are not associated with any particular type. As such, they
pose a problem for multifile builds because the type that we pick as the
owner ultimately gets to decide whether the methods will be compiled in
the current compilation.

On Project N, this is solved by dumping these methods into a compiler
generated assembly that is always part of the compilation. I'm mirroring
the same approach here, except the assembly is in memory only.

As part of fixing this, I had to split the
`MultiFileCompilationModuleGroup` into one that represents a production
build, and another one for the consumption build. This is something we
will need for #2282 too.

We might need a special rule to also COMDAT fold these.