Wrap the call to Main() method by another method (#2310)
This is necessary to protect the startup code infrastructure from
situation when the owning type of the main method cannot be loaded, and
codegen is instructed to generate a throwing body. Without wrapping,
this behavior would result in replacing the entire startup code sequence
with a throwing body, causing us to enter managed environment without it
being fully initialized. (In particular, the unhandled exception
experience won't be initialized, making this difficult to diagnose.)