Refactor PInvokeILEmitter for external configuration
The main motivation for this change is to allow multi-module builds to
demand lazy pinvokes when building libraries that would otherwise
generate eager pinvokes to OS-specific APIs. For example
`RoActivateInstance` only exists on Windows.

Add `PInvokeILEmitterConfiguration` class to capture externally
specified configuration, in this case to force lazy pinvokes when
compiling all code in an assembly into a library.  Move the static
native method counter out into `PInvokeILEmitterConfiguration` so its
lifetime is tied to the compilation.

Add `PInvokeILProvider` class which abstracts the `PInvokeILEmitter` and
`PInvokeILEmitterConfiguration` from `ILProvider`.  A
`PInvokeILProvider` instance is provided to `ILProvider` and will
eventually be switchable with something more functional than the current
PInvoke IL emitter.