Allow registration of multiple dispatch proxy generated assemblies.
With the incremental build, we need to be able to generate dispatch proxies per-assembly. As such we will end up with multiple dispatch proxy assemblies which all need to be able to register themselves at runtime, so that we can find the proxies when needed.
This change modifies the registration method to allow multiple calls to it and it concatenates the registered entries. Note that there's no need to deal with potential duplicates. The assumption is that the generated proxies are correct, so if we have two copies of the same proxy, those proxies will be identical in behavior. As such it doesn't matter which one will be used at runtime. So it's OK to leave both of them in the table. The lookup algorithm will pick the first one which matches the request, and since all of them should be identical in behavior it should work (tested that it really does work like this).

[tfs-changeset: 1642225]