Work necessary to enable RuntimeMethodHandles and RuntimeFieldHandles on CoreRT.
Today, these handles have a single IntPtr field, which points to some location in the module, corresponding to the beginning of the native signature of a method/field. At runtime, when we need to parse the signature of a method/field, we use the value of that signature pointer to figure out which module the RMH/RFH belongs to (RuntimeAugments.GetModuleFromPointer), load the NativeLayoutInfo blob from that module, and subtract the blob's address from the pointer address to get an offset into the native blob, to start the parsing.

This doesn't work well in CoreRT. In CoreRT, the plan is to have RuntimeMethodHandles and RuntimeFieldHandles contain a single pointer to a struct of 2 IntPtr pais: first IntPtr pointing at the TypeManager indirection cell, and second IntPtr containing the offset into the native layout.

[tfs-changeset: 1642745]