Add custom attribute to force static PInvoke resolution
Add `PInvokeStaticResolutionAttribute` which overrides the logic that
decides whether the PInvoke will be resolved by the compiler or lazily
at runtime. For multi-module compilation it is preferable to have
PInvokes all resolved lazily since we compile everything and there may
be unresolvable (but un-used) PInvokes.  However some PInvokes occur
early enough on the startup path that we can't do lazy resolution yet.
For example, calling LoadLibraryEx causes an infinite loop.