CR Feedback
Remove cctors under #if CORERT from types we want to initialize in an explicit order. Add a ILT_cctor method to those types. This matches the name used in .NET Native as closely as allowed in C# (it's actually $ILT$_cctor in .NET Native).
Replace the `ExplicitClassConstructors` class, which had a list of assemblies and types with eager cctors. Instead instruduce the `ModuleClassConstructors` class which has a list of assemblies with module-level initialization code to run. Each such assembly has a special type which calls the various ILT_cctors in that assembly. Change the lookup policy here to be non-fatal.
Use a wildcard for the class library so that if a new class library is introduced, it can provide the appropriate module initializer type and automatically get initialized.
Remove some old code in `CppWriter` that stubbed out the cctor in `ReflectionExecution`. Instead that is specified in the configuration for System.Private.Reflection.Execution in `ModuleClassConstructors`.
Move the statics in `ClassConstructorRunner+Cctor` up into the nesting class so we need one fewer eager cctors.