Adding ISerializable to System.Delegate/System.MulticastDelegate
Doing this now because the lack of it is forcing us to
spam apicompat baselines with exemptions for every
delegate type in the framework.

System.Delegate.GetObjectData() just throws
NotSupportedException so that's easy to implement now.

System.MulticastDelegate.GetObjectData() is more
complicated and probably needs MethodInfo
serialization to work first. So this is a stub.