Update default clang version to 3.9 and require xcode 8
This change updates the default clang version used to build CoreRT to the
latest stable version 3.9. It also removes the version 3.5 from the supported
ones. The 3.5 has a bug preventing us from using thread_local C++ feature.

Due to the same reason, the minimum supported xcode version on OSX is 8, so
the prerequisities document is updated accordingly.