Always generate static bases for library modules
Fixes issue https://github.com/dotnet/corert/issues/2568 where static
fields that aren't referenced anywhere in a module compiled as a library
won't get their static base generated in the image. This causes link
failures when an app is built against the library.

Added a test that is compiled as two separate modules and linked
together.