Fix stack overflow in execution of class constructors
The System.Threading.Lock class has a static field marked with the
PreInitialized attribute. The CoreRT compiler does not support this
attribute yet. This static field is accessed in the
Lock.TryAcquireContended method.

The ClassConstructorRunner class used the Lock type for internal
synchronization. As a result, we can get into an infinite recursion if
there is contention the ClassConstructorRunners lock before the cctor of
the Lock type had a change to execute.

This change fixes the issue by changing the Lock class to use Library
Initializers.

Note: this is issue was the root cause of the failures in the
BasicThreading test that we observed a couple times in the CI.