Fix Generics test
The fact we could reflection enable two instantiations of a method
relied on pure luck on which invoke stub we choose to invoke the method.
If we chose the one for `object`, we could use it to invoke the method
instantiated over `string`. We ran out of luck in postcheckin.

I'm fixning the test to only have one generic method instantiation to
make it pass again to keep coverage for the other interesting cases.

Making this actually work depends on the "Use type loader to build
invoke stub dictionaries" workitem. Requires type loader.