Fix a TODO in generating reflection invoke stubs (#2614)
I couldn't implement this at the point I added the TODO because we had
no way to access parameter metadata in the type system. We have it now.

I also had to make a few tweaks:
* Propagate parameter metadata to methods on instantiated types and
instantiated methods (doesn't make sense for interop's purposes, but
interop is not the only consumer)
* Empty array is a more convenient default value for the list

This has no test because this will only be testable once we start
instantiating the invoke stubs using the runtime typeloader.