Enable Enum-related apis for non-runtime Types.
This is a straight copy of the CoreClr code.

While it was tempting to create a third
subclass of EnumInfo and route it through
the regular Enum code for runtime types,
the runtime and default behaviors already
diverge in some weird edge cases.

Since third-party Type objects can
vary widely in quality and breadth
of implementation, it's probably simpler
to keep the code the same as CoreClr
and hope the reducer can pitch it out
for normal apps.