These are the binder and runtime changes to map IP values in cold code to the entry point of the hot code for methods that were hot/cold split.
The layout of this table uses the existing logic that maps IP values efficiently to integer indices, and it also exploits the fact that the cold code blocks are generally grouped together and thus have a contiguous group of method indices. The exception from is rule is the ARM processor where we may have multiple subsections of hot and cold code blocks.

Thus, the table contains the following items, in order:
 - the number of subsections (1 on x86 and amd64, bigger on ARM for modules with more than 16 MB of code)
 - an array containing the number of hot and cold method blocks for each subsection
 - an array containing the *hot* entry point RVA for each *cold* block

This layout combines fast lookup with modest space requirements (essentially a baseline of 12 bytes plus 4 bytes per method that was actually hot/cold split).

The encoding logic lives in zapcode.cpp and constructs the table described above.

The decoding logic lives in module.cpp and changes the semantics of GetMethodStartAddress to always return the start of the hot cold block if the IP happens to be in the cold block. I believe this change in semantics is general goodness and grepping through the source base have not found a place where this could cause trouble.

The cold-hot mapping logic is currently not active when DACCESS_COMPILE is defined - I wonder if it should be.

[tfs-changeset: 1647463]