Allow invalid binaries as input files to ILC
This is a hopefully short-term hack which allows files specified as
inputs to ILC to be actually native binaries that we ignore and proceed
with compilation.  The CoreCLR drops we test against contain a mixture
of managed and native binaries in test folders in some cases.

Also, add a quality of life switch to runtest.cmd (/CoreCLRSingleTest)
that allows specifying a single CoreCLR test to run by executable path.
This downloads the test blob and sets the correct environment variables
first.