Use a better mangled name for unbox+instantiating thunks
There was a bunch of code that I wrote to generate a name for
unbox+instantiating stubs that `// Mangle the namespace in the hopes
that it won't conflict with anything else.`

Turns out this is really hard because we need to also account for e.g.
method overloads and having to append suffixes that disambiguate between
them. There's one place in the compiler that can do this: the name
mangler.

What we really want is something like `"__unbox_" +
MangledNameOfTheInstanceMethod`.

Since the compiler type system context (which is where we make the
thunks) doesn't have a very convenient access to the name mangler (and
it would likely be a layering violation anyway), I explored two options:

* A secret interface (`IManglableMethod` that a `MethodDesc` could
implement) to provide custom mangling rules for the name mangler, or
* A new node type deriving from `MethodCodeNode` that has the name
mangling rule built in.

I chose the latter because it's simpler.