Add a rare flag to indicate EEType represents an abstract class
This is needed to support RuntimeHelpers.GetUninitializedObject - we need to prevent allocating abstract classes without triggering a `MissingMetadataException` (can't fall back to reflection metadata for this). I can also see how we could use this to give a better exception message from `Activator.CreateInstace<T>` (for the case of T being abstract) to improve CLR compatibility.

This assumes that once Project X is done, we'll have a discussion about what is and what isn't part of the EEType (what goes into separate hash tables). Maybe this is one of the things that should be in a side hash table.

[tfs-changeset: 1648986]