Move away from a static NameMangler (#2835)
Having a static NameMangler makes it an easy target for misuse and layering violations. Violating layering shouldn't be as easy as writing a one-liner that is easy to miss in a code review.

Having NameMangler not be static is also a requirement for making the code that uses it unit testable.