Add type loader support for building ByRef types
The CoreRT compiler and runtime have been able to handle this for a
while. The only thing missing (that required us to throw from a place)
was the type loader support.

This adds support for building ByRef types at runtime. I'm not
introducing a native layout representation though, so things like `T&`
won't compile still.