Emit all methods / data to Comdat sections
* When linking against a library, if the linker needs any symbols from
an object file in that library, it links in the entire library's
sections. This bloats the final binary image. When opt:ref is used (in
release mode), unreferenced symbols can be discarded if they are in
their own Comdat section.

* Switch to using Comdats for all ISymbolNodes in multi-module mode.

* This change is sufficient to collapse multi-module binaries down to
around 7MB (from 19MB) if metadata is disabled. In the presence of
metadata, all types and methods end up rooted.

* Remove the IsDelegate exclusion in library mode. We can handle rooting
delegates now and we were missing delegate types with metadata disabled.

* ObjectNode.IsShareable could strictly be removed at this point. I want
to let this bake a few days to be sure before doing that work though. We
may still want some of this logic, too; now everything is a global
mergeable comdat. The next step is to properly categorize symbols for
global / private linkage.