Fix casting arrays to non-variant interfaces
The runtime has a special set of rules for casting array types. Array
can be cast to various non-variant generic interfaces without triggering
an exception. Note this does not violate type safety because the safety
is enforced at the array store operation level.

The fix is twofold:
* Dependency tracking: methods implementing generic interfaces on arrays
should be always considered used (same as we do for variant interfaces).
This has a pretty negative size on disk impact. The work to see if we
can do something about it without adversely affecting compilation
throughput is tracked in #1198.
* EEType emission - this is a separate issue, but related: generic
interface types that can potentially be implemented by an array need to
have the GenericVariance flag set (even if all the parameters turn out
to be non-variant). This is to make casting take a slower path at
runtime that can handle special array casts.

There is one more issue that I discovered while looking at this in
`CanCastTo`. I'm not fixing it now. Tracked by #2871.