Generate method body for the delegate Invoke method (#2877)
We didn't really need this before because RyuJIT understands `Invoke`
method calls and inlines them. This is only to cover the case of `ldftn`
of the `Invoke` method and reflection-invoking the invoke method.

We could possibly also implement the body as `this.Invoke(args...)` (and
have RyuJIT expand that) but this has the benefit of also supporting
codegens that don't expand `Invoke`. On the other hand this has the
drawback of emitting a `calli` that will get the (unnecessary) fat
function pointer treatment.