Consider types loaded with LDTOKEN "constructed"
In Project N NUTC compiler, these types are considered "visible by
reflection" and as such, all bets as to what the code is going to do
with it are off. Whole program analysis could potentially optimize which
ones really need to be considered constructed (e.g. we often take an
EEType only to compare it with some other EEType), but that's a possible
optimization for the future.